audio.1.sg: +PUMA *SA. Ripper compat now official paradigm.
audio.1.sg:
 - Signature-enabled Sonic Arranger .SA midly tightened
 - PumaTracker .PUMA added in full

An ideological shift in my work has gained shape from now on: I'll start adding and/or improving the detectors in such a way that they will report the calculated size (in variable "sz") if worth the effort, and usually it is. Scripts reporting these can be converted to become parts of resource ripper software with minimal effort. We've had this before, but now it's official!

A lot of the new code is going to be parroting not only the code of ProWizard but also that of UADE's players' Motorola 68k assembly sources (the obscure but capable EagleRipper, in short), or something present that wasn't included yet, or something we'll personally figure out.