audio.1.sg: malformed reform, *ORG *MT2 *MMD? *DW *MED
This commit commodifies the way in which malformed files can be reported, and all malformed file reports in audio.1.sg are changed accordingly. See the update in TODO.md for an implementation recipe. (The necessary "read" and "_init" are attached but the recentmost DiE should have it too)

audio.1.sg:
 - Organya .ORG info-ed & ripper-ready
 - .MT2 finalised & ripper-ready
 - .DW completely switched to the way Nostalgic Player detects it, and even more info-ing added from it; that means SFX aren't detected as well, but also nothing even plays them...
 - (Octa)MED .MED/.MMD0~3 well-remade, .MMD ripper-ready and MED WIP